Playwright-cli/pyproject.toml
[project]
name = "feedback-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.3",
    "beautifulsoup4>=4.14.3",
    "html2text>=2025.4.15",
    "markdownify>=1.2.2",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.1",
]

